Create `login` feature

+++ app/Http/Controllers/DataMahasiswaController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Matakuliah;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class DataMahasiswaController extends Controller
 {
@@ -12,18 +13,22 @@ class DataMahasiswaController extends Controller
      */
     public function index()
     {
-        $data = [
-            'nim' => '200170120',
-            'nama' => 'Hawari Muflih Munte',
-            'prodi' => 'Teknik Informatika',
-            'jurusan' => 'Teknik Informatika'
-        ];
-        $db = Matakuliah::all();
+        if (Auth::check()) {
+            $data = [
+                'nim' => '200170120',
+                'nama' => 'Hawari Muflih Munte',
+                'prodi' => 'Teknik Informatika',
+                'jurusan' => 'Teknik Informatika'
+            ];
+            $db = Matakuliah::all();
 
-        return view('data_mahasiswa', [
-            'data_mahasiswa' => $data,
-            'matakuliah' => $db
-        ]);
+            return view('data_mahasiswa', [
+                'data_mahasiswa' => $data,
+                'matakuliah' => $db
+            ]);
+        }
+
+        return redirect()->route('login');
     }
 
     /**
@@ -31,7 +36,11 @@ class DataMahasiswaController extends Controller
      */
     public function create()
     {
-        return view('data_mahasiswa_create');
+        if (Auth::check()) {
+            return view('data_mahasiswa_create');
+        }
+
+        return redirect()->route('login');
     }
 
     /**
@@ -39,21 +48,25 @@ class DataMahasiswaController extends Controller
      */
     public function store(Request $request)
     {
-        $matakuliah = Matakuliah::create([
-            'kode' => $request->kode,
-            'nama' => $request->nama,
-            'sks' => (int)$request->sks,
-            'kelas' => $request->kelas,
-            'dosen' => $request->dosen
-        ]);
+        if (Auth::check()) {
+            $matakuliah = Matakuliah::create([
+                'kode' => $request->kode,
+                'nama' => $request->nama,
+                'sks' => (int)$request->sks,
+                'kelas' => $request->kelas,
+                'dosen' => $request->dosen
+            ]);
 
-        if ($matakuliah->exists()) {
-            session()->flash('message_create_success', 'Data berhasil ditambah');
-        } else {
-            session()->flash('message_create_failed', 'Data gagal ditambah. Terjadi kesalahan!');
+            if ($matakuliah->exists()) {
+                session()->flash('message_create_success', 'Data berhasil ditambah');
+            } else {
+                session()->flash('message_create_failed', 'Data gagal ditambah. Terjadi kesalahan!');
+            }
+
+            return redirect()->route('data-mahasiswa.index');
         }
 
-        return redirect()->route('data-mahasiswa.index');
+        return redirect()->route('login');
     }
 
     /**
@@ -61,15 +74,19 @@ class DataMahasiswaController extends Controller
      */
     public function show(string $id)
     {
-        $matakuliah = Matakuliah::find((int)$id);
+        if (Auth::check()) {
+            $matakuliah = Matakuliah::find((int)$id);
 
-        if ($matakuliah->exists()) {
-            session()->flash('message_show_success', 'Data berhasil ditemukan');
-        } else {
-            session()->flash('message_show_failed', 'Data gagal ditemukan. Terjadi kesalahan!');
+            if ($matakuliah->exists()) {
+                session()->flash('message_show_success', 'Data berhasil ditemukan');
+            } else {
+                session()->flash('message_show_failed', 'Data gagal ditemukan. Terjadi kesalahan!');
+            }
+
+            return view('data_mahasiswa_show', ['matakuliah' => $matakuliah]);
         }
 
-        return view('data_mahasiswa_show', ['matakuliah' => $matakuliah]);
+        return redirect()->route('login');
     }
 
     /**
@@ -77,15 +94,19 @@ class DataMahasiswaController extends Controller
      */
     public function edit(string $id)
     {
-        $matakuliah = Matakuliah::find((int)$id);
+        if (Auth::check()) {
+            $matakuliah = Matakuliah::find((int)$id);
 
-        if ($matakuliah->exists()) {
-            session()->flash('message_edit_success', 'Data berhasil ditemukan');
-        } else {
-            session()->flash('message_edit_failed', 'Data gagal ditemukan. Terjadi kesalahan!');
+            if ($matakuliah->exists()) {
+                session()->flash('message_edit_success', 'Data berhasil ditemukan');
+            } else {
+                session()->flash('message_edit_failed', 'Data gagal ditemukan. Terjadi kesalahan!');
+            }
+
+            return view('data_mahasiswa_edit', ['matakuliah' => $matakuliah]);
         }
 
-        return view('data_mahasiswa_edit', ['matakuliah' => $matakuliah]);
+        return redirect()->route('login');
     }
 
     /**
@@ -93,23 +114,27 @@ class DataMahasiswaController extends Controller
      */
     public function update(Request $request, string $id)
     {
-        $matakuliah = Matakuliah::find((int)$id);
+        if (Auth::check()) {
+            $matakuliah = Matakuliah::find((int)$id);
 
-        $matakuliah->update([
-            'kode' => $request->kode,
-            'nama' => $request->nama,
-            'sks' => (int)$request->sks,
-            'kelas' => $request->kelas,
-            'dosen' => $request->dosen
-        ]);
+            $matakuliah->update([
+                'kode' => $request->kode,
+                'nama' => $request->nama,
+                'sks' => (int)$request->sks,
+                'kelas' => $request->kelas,
+                'dosen' => $request->dosen
+            ]);
 
-        if ($matakuliah->exists()) {
-            session()->flash('message_update_success', 'Data berhasil diubah');
-        } else {
-            session()->flash('message_update_failed', 'Data gagal diubah');
+            if ($matakuliah->exists()) {
+                session()->flash('message_update_success', 'Data berhasil diubah');
+            } else {
+                session()->flash('message_update_failed', 'Data gagal diubah');
+            }
+
+            return redirect()->route('data-mahasiswa.index');
         }
 
-        return redirect()->route('data-mahasiswa.index');
+        return redirect()->route('login');
     }
 
     /**
@@ -117,14 +142,18 @@ class DataMahasiswaController extends Controller
      */
     public function destroy(string $id)
     {
-        $rowsAffected = Matakuliah::destroy((int)$id);
+        if (Auth::check()) {
+            $rowsAffected = Matakuliah::destroy((int)$id);
 
-        if ($rowsAffected > 0) {
-            session()->flash('message_delete_success', 'Data berhasil dihapus');
-        } else {
-            session()->flash('message_delete_failed', 'Data gagal dihapus');
+            if ($rowsAffected > 0) {
+                session()->flash('message_delete_success', 'Data berhasil dihapus');
+            } else {
+                session()->flash('message_delete_failed', 'Data gagal dihapus');
+            }
+
+            return redirect()->route('data-mahasiswa.index');
         }
 
-        return redirect()->route('data-mahasiswa.index');
+        return redirect()->route('login');
     }
 }
